modules/location-producer/requirements.txt
kafka-python==2.0.2
grpcio-tools==1.42.0
grpcio==1.42.0